requirements.txt
streamlit
torch
transformers
Pillow
gradio